refactor(context): extract budget parser module
